# remove burstiness in latency tests