--- a/akka-remote/src/main/scala/akka/remote/artery/AeronSink.scala
+++ b/akka-remote/src/main/scala/akka/remote/artery/AeronSink.scala
@@ -5,7 +5,6 @@ package akka.remote.artery
 
 import java.nio.ByteBuffer
 import java.util.concurrent.atomic.AtomicInteger
-
 import scala.annotation.tailrec
 import scala.concurrent.Future
 import scala.concurrent.Promise
@@ -14,7 +13,6 @@ import scala.util.Failure
 import scala.util.Success
 import scala.util.Try
 import scala.util.control.NoStackTrace
-
 import akka.Done
 import akka.stream.Attributes
 import akka.stream.Inlet
@@ -26,6 +24,7 @@ import akka.stream.stage.InHandler
 import io.aeron.Aeron
 import io.aeron.Publication
 import org.agrona.concurrent.UnsafeBuffer
+import org.agrona.hints.ThreadHints
 
 object AeronSink {
 
@@ -95,15 +94,17 @@ class AeronSink(
 
       private var completedValue: Try[Done] = Success(Done)
 
-      // FIXME measure and adjust with IdleCpuLevel
-      private val spinning = 1000
+      // spin between 2 to 20 depending on idleCpuLevel
+      private val spinning = 2 * taskRunner.idleCpuLevel
       private var backoffCount = spinning
       private var lastMsgSize = 0
-      private val offerTask = new OfferTask(pub, null, lastMsgSize, getAsyncCallback(_ ⇒ onOfferSuccess()),
+      private val offerTask = new OfferTask(pub, null, lastMsgSize, getAsyncCallback(_ ⇒ taskOnOfferSuccess()),
         giveUpSendAfter, getAsyncCallback(_ ⇒ onGiveUp()))
       private val addOfferTask: Add = Add(offerTask)
 
       private var offerTaskInProgress = false
+      private var delegateTaskStartTime = 0L
+      private var countBeforeDelegate = 0L
 
       private val channelMetadata = channel.getBytes("US-ASCII")
 
@@ -135,10 +136,10 @@ class AeronSink(
       @tailrec private def publish(): Unit = {
         val result = pub.offer(envelopeInFlight.aeronBuffer, 0, lastMsgSize)
         // FIXME handle Publication.CLOSED
-        // TODO the backoff strategy should be measured and tuned
         if (result < 0) {
           backoffCount -= 1
           if (backoffCount > 0) {
+            ThreadHints.onSpinWait()
             publish() // recursive
           } else {
             // delegate backoff to shared TaskRunner
@@ -146,12 +147,20 @@ class AeronSink(
             // visibility of these assignments are ensured by adding the task to the command queue
             offerTask.buffer = envelopeInFlight.aeronBuffer
             offerTask.msgSize = lastMsgSize
+            delegateTaskStartTime = System.nanoTime()
             taskRunner.command(addOfferTask)
-            flightRecorder.hiFreq(AeronSink_DelegateToTaskRunner, lastMsgSize)
+            flightRecorder.hiFreq(AeronSink_DelegateToTaskRunner, countBeforeDelegate)
           }
         } else {
+          countBeforeDelegate += 1
           onOfferSuccess()
         }
+      }
+
+      private def taskOnOfferSuccess(): Unit = {
+        countBeforeDelegate = 0
+        flightRecorder.hiFreq(AeronSink_ReturnFromTaskRunner, System.nanoTime() - delegateTaskStartTime)
+        onOfferSuccess()
       }
 
       private def onOfferSuccess(): Unit = {