--- a/akka-remote/src/main/scala/akka/remote/artery/AeronSource.scala
+++ b/akka-remote/src/main/scala/akka/remote/artery/AeronSource.scala
@@ -4,10 +4,8 @@
 package akka.remote.artery
 
 import java.util.concurrent.TimeUnit
-
 import scala.annotation.tailrec
 import scala.concurrent.duration._
-
 import akka.stream.Attributes
 import akka.stream.Outlet
 import akka.stream.SourceShape
@@ -22,6 +20,7 @@ import io.aeron.logbuffer.FragmentHandler
 import io.aeron.logbuffer.Header
 import org.agrona.DirectBuffer
 import org.agrona.concurrent.BackoffIdleStrategy
+import org.agrona.hints.ThreadHints
 
 object AeronSource {
 
@@ -80,18 +79,15 @@ class AeronSource(
     new GraphStageLogic(shape) with OutHandler {
 
       private val sub = aeron.addSubscription(channel, streamId)
-      // FIXME measure and adjust with IdleCpuLevel
-      private val spinning = 1000
-      private val yielding = 0
-      private val parking = 0
-      private val idleStrategy = new BackoffIdleStrategy(
-        spinning, yielding, TimeUnit.MICROSECONDS.toNanos(1), TimeUnit.MICROSECONDS.toNanos(100))
-      private val idleStrategyRetries = spinning + yielding + parking
-      private var backoffCount = idleStrategyRetries
+      // spin between 100 to 10000 depending on idleCpuLevel
+      private val spinning = 1100 * taskRunner.idleCpuLevel - 1000
+      private var backoffCount = spinning
+      private var delegateTaskStartTime = 0L
+      private var countBeforeDelegate = 0L
 
       // the fragmentHandler is called from `poll` in same thread, i.e. no async callback is needed
       private val messageHandler = new MessageHandler(pool)
-      private val addPollTask: Add = Add(pollTask(sub, messageHandler, getAsyncCallback(onMessage)))
+      private val addPollTask: Add = Add(pollTask(sub, messageHandler, getAsyncCallback(taskOnMessage)))
 
       private val channelMetadata = channel.getBytes("US-ASCII")
 
@@ -107,8 +103,7 @@ class AeronSource(
 
       // OutHandler
       override def onPull(): Unit = {
-        idleStrategy.reset()
-        backoffCount = idleStrategyRetries
+        backoffCount = spinning
         subscriberLoop()
       }
 
@@ -118,22 +113,29 @@ class AeronSource(
         val msg = messageHandler.messageReceived
         messageHandler.reset() // for GC
         if (fragmentsRead > 0) {
+          countBeforeDelegate += 1
           if (msg ne null)
             onMessage(msg)
           else
             subscriberLoop() // recursive, read more fragments
         } else {
-          // TODO the backoff strategy should be measured and tuned
           backoffCount -= 1
           if (backoffCount > 0) {
-            idleStrategy.idle()
+            ThreadHints.onSpinWait()
             subscriberLoop() // recursive
           } else {
             // delegate backoff to shared TaskRunner
-            flightRecorder.hiFreq(AeronSource_DelegateToTaskRunner, 0)
+            flightRecorder.hiFreq(AeronSource_DelegateToTaskRunner, countBeforeDelegate)
+            delegateTaskStartTime = System.nanoTime()
             taskRunner.command(addPollTask)
           }
         }
+      }
+
+      private def taskOnMessage(data: EnvelopeBuffer): Unit = {
+        countBeforeDelegate = 0
+        flightRecorder.hiFreq(AeronSource_ReturnFromTaskRunner, System.nanoTime() - delegateTaskStartTime)
+        onMessage(data)
       }
 
       private def onMessage(data: EnvelopeBuffer): Unit = {