--- a/akka-remote/src/main/scala/akka/remote/artery/FlightRecorderEvents.scala
+++ b/akka-remote/src/main/scala/akka/remote/artery/FlightRecorderEvents.scala
@@ -28,12 +28,14 @@ object FlightRecorderEvents {
   val AeronSink_EnvelopeOffered = 18
   val AeronSink_GaveUpEnvelope = 19
   val AeronSink_DelegateToTaskRunner = 20
+  val AeronSink_ReturnFromTaskRunner = 21
 
   // Aeron Source events
-  val AeronSource_Started = 21
-  val AeronSource_Stopped = 22
-  val AeronSource_Received = 23
-  val AeronSource_DelegateToTaskRunner = 24
+  val AeronSource_Started = 22
+  val AeronSource_Stopped = 23
+  val AeronSource_Received = 24
+  val AeronSource_DelegateToTaskRunner = 25
+  val AeronSource_ReturnFromTaskRunner = 26
 
   // Compression events
   val Compression_CompressedActorRef = 25