--- a/akka-remote/src/test/resources/aeron.properties
+++ b/akka-remote/src/test/resources/aeron.properties
@@ -11,6 +11,11 @@ agrona.disable.bounds.checks=true
 
 aeron.threading.mode=SHARED_NETWORK
 
+# low latency settings
+#aeron.threading.mode=DEDICATED
+#aeron.sender.idle.strategy=org.agrona.concurrent.BusySpinIdleStrategy
+#aeron.receiver.idle.strategy=org.agrona.concurrent.BusySpinIdleStrategy
+
 # use same director in akka.remote.artery.advanced.aeron-dir config 
 # of the Akka application 
 aeron.dir=target/aeron