* throttle generates bursts but for fair latency tests
  we want the messages to be spread uniformly