@patriknw 2016-06-15
Before: === Latency rate-10000-size-100: RTT 50%ile: 153 µs, 90%ile: 225 µs, 99%ile: 910 µs, rate: 9,573 msg/s

After: === Latency rate-10000-size-100: RTT 50%ile: 1108 µs, 90%ile: 1544 µs, 99%ile: 2286 µs,

Histogram before and after at 10,000 msg/s

![histogram 1](https://cloud.githubusercontent.com/assets/336161/16083234/8f4290d2-3314-11e6-849e-24b2cd75fc4b.png)

![histogram 3](https://cloud.githubusercontent.com/assets/336161/16083449/4714b44c-3315-11e6-8b10-716552de14e6.png)