@patriknw 2016-06-19
@drewhk I have added flight recorder metrics events for the task runner. I think that will be useful when tuning the idle strategies. see 735532e