@patriknw 2016-06-22
I have also tried to tune the `idle-cpu-level`, see aca4ebd

Here are the results of the `LatencySpec` running warmup, rate 100, 1000 and 10000 msg/s at the different levels.

The test was run with embedded media driver, i.e. 2 media drivers. Less cpu would have been used if I used one shared external media driver.

## Level 1

![level1](https://cloud.githubusercontent.com/assets/336161/16261128/bbccb078-386a-11e6-8c9d-64ee3eff902e.png)

<img width="234" alt="cpu-level1" src="https://cloud.githubusercontent.com/assets/336161/16261173/e2389a4c-386a-11e6-93c5-e488bdee1958.png">

## Level 3

![level3](https://cloud.githubusercontent.com/assets/336161/16261194/f89cc8f8-386a-11e6-863a-e9bcd6753aa7.png)

<img width="215" alt="cpu-level3" src="https://cloud.githubusercontent.com/assets/336161/16261197/fe8dede6-386a-11e6-971a-99cd09f75b2e.png">

## Level 5

![level5](https://cloud.githubusercontent.com/assets/336161/16261210/0e05203c-386b-11e6-8e6e-5a54a230fca8.png)

<img width="210" alt="cpu-level5" src="https://cloud.githubusercontent.com/assets/336161/16261221/1796f472-386b-11e6-8f56-1398af008b8f.png">

## Level 7 (best on my machine)

![level7](https://cloud.githubusercontent.com/assets/336161/16261266/351e7ae2-386b-11e6-9705-dce9370027a6.png)

<img width="213" alt="cpu-level7" src="https://cloud.githubusercontent.com/assets/336161/16261282/412b0a12-386b-11e6-9353-8876c949a42f.png">

## Level 9

![level9](https://cloud.githubusercontent.com/assets/336161/16261353/82d15890-386b-11e6-84ca-275080b98b8e.png)

<img width="216" alt="cpu-level9" src="https://cloud.githubusercontent.com/assets/336161/16261364/887e3f10-386b-11e6-9c4b-a8e7c4c77f2e.png">

## Level 10 (busy spin)

![level10](https://cloud.githubusercontent.com/assets/336161/16261376/99592606-386b-11e6-9a23-027b61d39161.png)

<img width="242" alt="cpu-level10" src="https://cloud.githubusercontent.com/assets/336161/16261389/a27c799a-386b-11e6-8179-59d91bc6d3f4.png">

## Rate 100 msg/s

![rate100](https://cloud.githubusercontent.com/assets/336161/16261418/bc620cbc-386b-11e6-917b-b6d44fefab73.png)

## Rate 1000 msg/s

![rate1000](https://cloud.githubusercontent.com/assets/336161/16261427/c646f6b6-386b-11e6-8896-597fc491742e.png)

## Rate 10000 msg/s

![rate10000](https://cloud.githubusercontent.com/assets/336161/16261434/cfced3ac-386b-11e6-9c57-ee6de5793034.png)