@patriknw 2016-06-22
This is how level 10 looks like when using external media driver.

![level10ext](https://cloud.githubusercontent.com/assets/336161/16264450/08809114-387a-11e6-8c9f-9b0157f0e213.png)

<img width="265" alt="cpu-level10ext" src="https://cloud.githubusercontent.com/assets/336161/16264467/1e3c3b70-387a-11e6-8968-2b1bfc74a60d.png">