@patriknw 2016-06-22
This is how old remoting looks like.

![old](https://cloud.githubusercontent.com/assets/336161/16264505/54e622d0-387a-11e6-9936-b52b766c047b.png)

<img width="214" alt="cpu-old" src="https://cloud.githubusercontent.com/assets/336161/16264519/6350ce6a-387a-11e6-87f6-37f4dece28e0.png">

and comparing old vs level 7

![old-vs-level7](https://cloud.githubusercontent.com/assets/336161/16264510/59c2ec5c-387a-11e6-97eb-a3be93f914cb.png)