@patriknw 2016-06-22
and finally, this is what it looks like at the jenkins servers: https://jenkins.akka.io:8498/job/akka-artery-repeat/plot/LatencySpec/

Build 66-67 was `artery-dev` baseline.
Build 68-70 was when removing bursting in test.
Build >= 72 is this branch, i.e. with the tuning