@viktorklang 2016-06-22
Interesting!

Looks like some drastic improvements have been achieved.
What does throughput look like?

On Wed, Jun 22, 2016 at 1:15 PM, Patrik Nordwall <notifications@github.com>
wrote:

> and finally, this is what it looks like at the jenkins servers:
> https://jenkins.akka.io:8498/job/akka-artery-repeat/plot/LatencySpec/
>
> Build 66-67 was artery-dev baseline.
> Build 68-70 was when removing bursting in test.
> Build >= 72 is this branch, i.e. with the tuning
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/akka/akka/pull/20790#issuecomment-227713648>, or mute
> the thread
> <https://github.com/notifications/unsubscribe/AAAqd8l0iqwZEQrH_M5x4VQ5yktlPZg4ks5qORlAgaJpZM4I2Z2H>
> .
>



-- 
Cheers,
√