@patriknw 2016-06-22
Throughput has not changed by this. On my machine it's around 700,000 msg/s for the small 100 bytes message (which is actually > 200 bytes due to lack of header compression).

On EC2 we observed the following between two instances:
* 460,000 msg/s for message payload of 100 bytes
* 90 Mbytes/s for message payload of 10,000 bytes

However, we think it maxed at that because of 1 Gbit/s network. We need to try with 10Gbit.