@patriknw 2016-07-01
rebased